refactor: rename `@multimodal/agent-cli` to `@tarko/cli`
