+++ multimodal/tarko/cli/bin/cli.js
