+++ multimodal/tarko/cli/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@multimodal/agent-cli",
+  "name": "@tarko/cli",
   "version": "0.3.0-beta.1",
   "description": "A flexible CLI framework for building multimodal agent applications",
   "main": "dist/index.js",
