+++ multimodal/tarko/cli/src/config/paths.ts
