+++ multimodal/tarko/cli/src/core/cli.ts
