+++ multimodal/tarko/cli/src/core/commands/index.ts
