+++ multimodal/tarko/cli/src/core/commands/run.ts
