+++ multimodal/tarko/cli/src/core/commands/start.ts
