+++ multimodal/tarko/cli/src/core/index.ts
