+++ multimodal/tarko/cli/src/core/stdin.ts
