+++ multimodal/tarko/cli/src/utils/console-interceptor.ts
