+++ multimodal/tarko/cli/src/utils/logo.ts
