+++ multimodal/tarko/cli/tsconfig.json
